Create Suggestions
This is a really cool program... Really creative. While I love the idea, it's a little bit hard to follow. I would suggest first organizing the code into more coherent methods so that you have reasonable code formatting and differentiate different functions. Also, if you commented in your code to explain what's going on that would be IMMENSELY helpful!!!! 
Another thing I would say is that maybe in your description, or in your comments, you could explain your rating system. Since the system was designed by you, it would be cool to know what contributes to getting a higher or lower rating. 
If you are thinking about adding to this, a graph of comparisons would be super cool! And would be an opportunity to have a GUI of some sort :)
Overall this is really cool, just would suggest making it easier to follow!
-Serena